fix(ui): move updateNodeInternals from general node to handles
